refactor(loop): split agent loop prompt helpers
